sr_rabbit was downloading rabbitmqadmin to the current working directory on any invocation of sr_audit.  So every time sr_audit was run in a new directory, it would place rabbitmqadmin there, and then run it.   After the first invocation, that script is trusted.
For some reason, this was bombing... so I replaced this with logic that just expects rabbitmqadmin
to be in the path somewhere, rather than explicitly downloading and running it.

I found out why it was bombing later... the password had a semi-colon in it, but the password was not
quoted.  Added quotes and all is well.