The meaning of state variables changed, so changed their names.
unready is now used to store all events which have not yet been processed, rather than only those
for files that weren't readable.  so it's name was changed to 'new_events'

unready_tmp was the list of events being processed when working on the backlog.
Now it refers to events which are being processed, but it also can persist, in the case of the original
use of unready.  so changed name fo 'events_outstanding'.

new_events is now copied to events_outstanding evert 'sleep' interval, then reset to empty (because all events are now outstanding.)
Then follows a loop to process the events_outstanding.

if we succeed in processing all events_oustanding, then it should be empty.