C - bigger patch than would prefer:
symlinks now work.
API changed for post to add a struct stat as an (read-only) argument.
so if the file doesn't exist locally, just fill in the stat buffer appropriately.
renamed add_header to header_add
added   header_reset() to set table to empty.
there was a bug bisected to an earlier change, that had killed off all the headers, reverted.