Code being deleted: self.channel.basic_recover(requeue=True)
PAS: 2020, I read this: https://groups.google.com/forum/#!topic/rabbitmq-users/UFz59V2M3W8
summary:
basic_recover was intended to be called when a client is overwhelmed. It was found in
practice that when clients were overwhelmed, they were usually unable to issue
the call, so recovery is entirely broker side in later protocol versions.

Also read this: https://pypi.org/project/amqp/
The basic_recover (and async) call is deprecated in later versions of AMQP library. (0.9.1)

The channel was set up in the line above, there can never be any messages to requeue.
So... calling it when setting up the connection does nothing, at best.
It might cost some back and forth traffic.