self.hc.loop guard removed in Consumer.consume()
1. There does not seem to be any situation where one would call consume()
   and not want to loop until success.

2. A case was identified in testing where, if basic_get fails, but self.hc.loop is false,
   it will just loop forever tightly, without even a sleep.

3. There are no cases where self.hc.loop would be set to false in the actual code.
   (All code paths leading to consume will set it to True prior to calling it.)

From 2. The immediate question is:  are there likely to be transient errors
that occur in basic_get that would be remedied by calling it again.
Nothing other than a problem with the connection comes to mind.  To double check,
looked in code of one amqp library (python3-amqp) and the the only error it would
seem able to raise is "RecoverableConnectionError('connection already closed')"
which would require the connection to be re-opened, which is what the recovery
code guarded by self.hc.loop does.

so the null error handling in 2 is unlikely to help in any situation, and
the actually helpful behaviour is already present behind a guard.

The result is that an artificial failure is provoked in testing by having basic_get
fail, with self.hc.loop being false.

Given the above, it is more representative of the application to remove the guard.