Move from pyCodeStyle to flake8
pyCodeStyle dev says for a more feature rich linting experience to use
flake8 as a wrapper around pyCodeStyle. Updated the setup.cfg and linter
script to use it instead.

This essentially means we can now make use of the # NoQA string
(No Quality Assurance forces the linter to not report on certain errors)