Redis driver for Nodupe (#705)
* Added many more NoDupe tests

* Add final tests to nodupe

* Add nodupe sub-class tests

* Created nodupe_redis

Stripped a lot of the methods out, because they just aren't needed with Redis.

* First try at Redis Nodupe

* Unit test for nodupe_redis

* Merged nodupe derivekey tests into 1

Previously had a whole bunch of them, but since we're re-instanciating the message for each derivation, they can all go in one.

* Unit test for nodupe_redis

Had to make some changes to the module itself because some functionality wasn't working

* Create generic nodupe unit test

Trying to compare the functionality of the disk-, and redis-based nodupe classes is going to be hard. Most (all?) of the methods don't actually return anything, and if we're assuming to abstract the inner workings of each one, there's nothing to validate they're doing the same thing.

* Delete individual derivekey nodupe unit tests

* Add test docs on using VSCode

* Allow picking which driver to use for NoDupe

* Add missing config options for redis nodupe

* Fix cache size counting

I'm not sure how much any of this counting matters, or if it's just for logging purposes.
Either way.. it's better now.

* Fix broken nodupe.redis tests

* Change field delimiter in nodupe redis key

* Update redis key field delimiter in nodupe test

* Fixed mis-named config option

Nodupe Redis was using retryqueue server URL.
now it's not.

* Standardized some testing aspects

- Made the add_option method for options work properly
- Added PrettyPrinter code for debugging tests

* Clean-up test debugging code

* Make Redis Nodupe count better

It wasn't loading on_start, or clearning on_stop, so it was never accurate.
I still don't really think it's all that accurate but since it's just for logging purposes, it might not need to be better than it is.

* Disabled uneeded redis Retry test

* Update retry unit test dependencies

Used to have dependencies per method, but now the comparative tests depend on both driver tests.
This ensures that nothing gets missed.

* Comparative NoDupe unit test

Compares state/output of both Disk and Redis nodupe drivers against an expected value.

* Change integrity to identity, per #703 and #706

This might cause merge conflicts, but it's probably still worth changing here.

* Resolve config.py conflict

* Resolve second config.py conflict

---------

Co-authored-by: Peter Silva <peter@bsqt.homeip.net>