Update unit testing patterns (#731)
* Expand test debugging (pretty)

Pretty printing wasn't awesome, but now it's better.

* Move Options into a proper class

* Move message creation into method

This ensures that there's no collisions between tests if we modify properties/dict values

* Use deepcopy on WorkList creation in each test

This ensures that they're always unique, and have no collisions between tests

* Fix broken retry tests

* Add Note to top of retry "steps" test

It's disabled, but can serve as good documentation of how to do it.

* Capture test log existance, and assert against that

It used to assert inside the "if" statement, but that would only ever be True. This way, it would be possible for the logs to be missing, and the test will fail.

* Fix redisqueue cleanup test

No longer need to specify a fake server, becasue we're checking the keys a little more inteligently.

* Clean up commented code

* Expand test debugging (pretty)

Pretty printing wasn't awesome, but now it's better.

* Move Options into a proper class

* Move message creation into method

This ensures that there's no collisions between tests if we modify properties/dict values

* Use deepcopy on WorkList creation in each test

This ensures that they're always unique, and have no collisions between tests

* Fix broken retry tests

* Add Note to top of retry "steps" test

It's disabled, but can serve as good documentation of how to do it.

* Capture test log existance, and assert against that

It used to assert inside the "if" statement, but that would only ever be True. This way, it would be possible for the logs to be missing, and the test will fail.

* Fix redisqueue cleanup test

No longer need to specify a fake server, becasue we're checking the keys a little more inteligently.

* Clean up commented code