Unit tests for all Flowcb/Accept plugins (#757)
* Bunch more callback tests in flowcb/accept

* Bunch more callback tests in flowcb/accept

* Add more tests for flowcb/accept

* Fix tests for flowcb/accept/tohttp

THey were broken because of the 'set_notice' header being wrong

* More flowcb/accept tests!!

flowcb/accept/postoverride
flowcb/accept/tohttp
flowcb/accept/wmotypesuffix

* Remove set_notice items in message

That item was removed as part of writing unit tests

* Bunch more callback tests in flowcb/accept

* Add more tests for flowcb/accept

* Fix tests for flowcb/accept/tohttp

THey were broken because of the 'set_notice' header being wrong

* More flowcb/accept tests!!

flowcb/accept/postoverride
flowcb/accept/tohttp
flowcb/accept/wmotypesuffix

* Remove set_notice items in message

That item was removed as part of writing unit tests

* Bunch more callback tests in flowcb/accept

* Add more tests for flowcb/accept

* Fix tests for flowcb/accept/tohttp

THey were broken because of the 'set_notice' header being wrong

* More flowcb/accept tests!!

flowcb/accept/postoverride
flowcb/accept/tohttp
flowcb/accept/wmotypesuffix

* Remove set_notice items in message

That item was removed as part of writing unit tests

* Bunch more callback tests in flowcb/accept

* Remove set_notice items in message

That item was removed as part of writing unit tests

* More flowcb/accept tests

downloadbaseurl
renamedmf
tolocalfile

All with 100% coverage

* Yes more flowcb/accept tests

* Final set of flowcb/accept unit tests

* Bunch more callback tests in flowcb/accept

* Add more tests for flowcb/accept

* Fix tests for flowcb/accept/tohttp

THey were broken because of the 'set_notice' header being wrong

* More flowcb/accept tests!!

flowcb/accept/postoverride
flowcb/accept/tohttp
flowcb/accept/wmotypesuffix

* Remove set_notice items in message

That item was removed as part of writing unit tests

* Bunch more callback tests in flowcb/accept

* Remove set_notice items in message

That item was removed as part of writing unit tests

* More flowcb/accept tests

downloadbaseurl
renamedmf
tolocalfile

All with 100% coverage

* Yes more flowcb/accept tests

* Final set of flowcb/accept unit tests

* Final fixes for flowcb/accept plugins

Should have been part of the inital set of fixes, but they weren't known until the tests were fixed.. sort of circular problem.

* Update test config

There's a class we're testing that starts with 'Test', and it was throwing warnings for pytest. Updating the config to only look for classes that start with 'Test_*' now.

* Fixes to all flowcb/accept tests

These fixes reflect the changes that were done in #756
Some of them also are fixes to make sure these tests are successful when run as part of the whole suite (caplog message counts among them). Those assertions probably aren't all that purposeful, and checking that the desired logs are in the list is better, but here we are..