instance_pids if no pid file are not missing.
This code says the instances we find that don't have pids, are missing.
This happens when we read metrics, and then know what the pids of the
flows were when it was last running.
An instance pid that doesn't have a matching file is actually a stray.
I think this code pre-dates strays. anyways what was happenning:

sr3 stop xx   # xx has current metrics file.
sr3 sanity
  # instances of xx are detected as "missing" because metrics present.
  # sanity therefore starts up xx when it should not.