remove plugins that are useless and will not run.
the tolocal.py expects a message field "urlstr" to be present in the
message, which isn't there in v3. so it will just crash if you try to
use it.  We could fix it, but what does it do? It converts a web
url to a local file url...

We already have working tolocalfile.py which does that.  The only
"advantage" of tolocal is that it has a paired callback "tohttp.py"
but that just does what an ordinary *post_baseURl* setting will do
in sr3. so tohttp.py is useless, tofile.py doesn't work and there is
an existing working implementation anyways... so... bye...