do not mkdirs in _resolve.
This patch is a second try at fixing 1080. It looks a bit wider.
the _resolve() routine where the error occurs is supposed to be
bringing together the data collected by other routines, and making
correlations. It is odd that there are mkdirs in there. This has
the effect of creating state dirs, when doing *sr3 status* which
feels wrong.  so remove the mkdirs, and see what breaks.