suppress error message about missing subscriptions.json
The first time you run any configuration (or any existing configuration
from an older version) you will see unable to open error.  It's normal
for it not to exist on first startup (or first after upgrade.)
I just demoted it to debug.  Not sure if there are useful cases
of this error.