symlinks are always relative to location of file.
working on #1270.

The relative path conversion is actually different for rename vs. link.
the rename needs to take into account where in the tree it is,
but the link is always from the destination directory (where the link
is.)  So the processing of 'link' and 'rename' is actually different.

I need more testing, I think link is correct now.