changed logic for sr status disconnected & down.
the problem was that there were multiple conditions to check
to assign one of the new states and if the multiple conditions
did not result in "down" or "disconnected" state...  it would
skip over a number of checks for *idle* state.
So moved all the new states to a separate if
above the exising series.