review and improve error handling when invoking makedirs (#1365)
* review and improve error handling when invoking makedirs

If (usual case) there are multiple files in a directory,
then there is a race condition to create the directory, multiple
instances will see the directory is not there, and will run makedirs
to create it. the "loser" just won't create the directory... but with
unlink... the unlink fails (because directory exists.) and
then the operation fails. (not retried? FIXME?)

So I think it's better to leave that failure mode to humans.