add *template* to subscription, to be able to match old_subscriptions.
had the problem that if someone declares a second queue on the same
broker, when reading from the subscription, it would match on the
broker... so the two subscriptions would collapse into one.

To fix that, created a "template" (the value of the queueName option
as entered in the config) as an additional matching parameter.  Now
one can have multiple queues on the same broker without risk of
confusion.