add starting,stopping states, diable sanity then.
There is a new routine in sr.py, _tag_progress()...
it is used to manage creation of "disabled", "starting", "shutdown"
files. These files tag for the status and sanity commands that
a given configuration is not in a consistent state.

Adds "starting", and "shutdown" states to known states of GlobalState.
It is normal in these states for instance counts and running processes
to not match. when building the status the configs that are
in one of the above states are accumulated in *self.flux* .
if self.flux is non-empty, then sanity refuses to change anything.

Also tightened some timing around starting and stopping instances.