match statedir with config to stop false strays
Check that the statehost variable matches what it should be
when checking a given state directory.

In the dynamic flow test, the sarra/download_f20 tasks were getting
flagged as strays by sanity because in a different test, statehost
is on, and so the statehost state directory (which should contain
pid files) was being read making sanity start extra instances.