user mode systemd unit, with co-ordinated cgroups.
Add option setting cgroupOverride, example usage:

cgroupOverride /sys/fs/cgroup/user.slice/user-1000.slice/user@1000.service/app.slice/metpx-sr3.service

this means that when sr3 starts up a process, either by analyst intervention
or sr3 sanity, it will add the processes launched to this cgroup.
systemD uses this group, so when starting up entire thing with systemD it will see what is already
running.  so systemD view and sr3 view of which processes are running should be consistent now.

otoh... if systemd things the unit is stopped, it does not look at the cgroup, so it will not
see things running if they are started manually.  once it is started though, it does see
the ones manually started before.

~
~